Use the new exclude_from_vendor_snapshot property.
This allows the associated forked modules, which are
normally considered framework modules, and would be
included in the vendor snapshot, to be excluded from the
vendor snapshot.

Since these modules are forked, the vendor has assumed
responsibility for maintaining these modules. Adding the
exclude_from_vendor_snapshot property causes them to be
excluded from the vendor snapshot, and the modules will
be built from source when building the vendor image
against the VSDK.

Bug: 165705527
Test: build partner code using vsdk
Test: basic functional test
Change-Id: Ic21b72b942d145b53ee5c18c57847afe0dd2b9e6